requirements.txt
django
djangorestframework
djoser
djangorestframework-simplejwt
mysqlclient
channels
daphne